Improve performance of (large) directory listings: Avoid stat:ing each file many times.
The patch modifies rdesktop so that the g_notify_stamp is only set
when writing. Also, the stamp is not set before disk_create_notify(),
since this would mean that NotifyInfo would be called twice directly.

With this patch, the number of stat:s has dropped from 24 to 4, using
my tests.


git-svn-id: https://rdesktop.svn.sourceforge.net/svnroot/rdesktop/trunk/rdesktop@1401 9cebe5d2-de03-4e0a-9b96-fbf3d975a676